% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gfigpd}
\alias{gfigpd}
\title{Title}
\usage{
gfigpd(
  X,
  beta,
  i = floor(0.85 * length(X)),
  gamma.init = NA,
  sigma.init = NA,
  sd.gamma = NA,
  sd.sigma = NA,
  p1 = 0.9,
  p2 = 0.5,
  lambda1 = 2,
  lambda2 = 10,
  Jnumb = 50L,
  iter = 10000L,
  burnin = 2000L,
  thin = 6L,
  nchains = nthreads,
  nthreads = parallel::detectCores(),
  seeds = NULL
)
}
\arguments{
\item{X}{numeric vector of data}

\item{beta}{vector of probabilities corresponding to the quantiles to be 
estimated}

\item{i}{index for the initial threshold at the \code{X(i)} order statistic}

\item{gamma.init}{starting value for gamma in the MCMC}

\item{sigma.init}{starting value for sigma in the MCMC}

\item{sd.gamma}{standard deviation for the proposed gamma}

\item{sd.sigma}{standard deviation for the proposed sigma}

\item{p1}{probability that the MCMC will propose a new \code{(gamma,sigma)}; 
\code{(1-p1)} would be the probability that the MCMC chain will propose a 
new index for a new threshold}

\item{p2}{probability that the new index proposed will be larger than the 
current index}

\item{lambda1}{the small jump the index variable will make}

\item{lambda2}{the large jump the index variable will make; happens 1 of 
every 10 iterations}

\item{Jnumb}{number of subsamples that are taken from the Jacobian}

\item{iter}{number of iterations per chain (burnin excluded)}

\item{burnin}{number of the first MCMC iterations discarded}

\item{thin}{thinning number for the MCMC chain. (e.g. if it is 1 no iteration 
is skipped)}

\item{nchains}{number of MCMC chains to run}

\item{nthreads}{number of threads to run the chains in parallel}

\item{seeds}{the seeds used for the MCMC sampler; one seed per chain, or 
\code{NULL} to use random seeds}
}
\value{

}
\description{
Title
}
\examples{
data("rain", package = "ismev")
gf <- gfigpd(rain, beta = c(0.98, 0.99), iter = 3000L)
}
